services/web/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.4
gunicorn==20.1.0
PyYAML==6.0
dict2xml==1.7.2 